refactor: separate the site and the parser code
