+++ parser/.gitignore
@@ -1,0 +1,5 @@
+node_modules/
+lib/
+lib-esm/
+_bundles/
+.idea
