+++ parser/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "target": "es5",
+    "module": "commonjs",
+    "lib": ["es2015", "dom"],
+    "outDir": "lib",
+    "allowJs": true,
+    "sourceMap": true,
+    "declaration": true,
+    "strict": true,
+    "moduleResolution": "Node"
+  },
+  "exclude": [
+    "./webpack.config.js",
+    "node_modules",
+    "lib",
+    "lib-esm",
+    "_bundles"
+  ]
+}
