+++ site/src/components/Report/Article/index.tsx
