+++ site/src/components/Report/DepartmentsChart/index.tsx
