+++ site/src/components/Report/SeriesChart/SeriesChart.module.scss
