+++ site/src/components/Report/SeriesChart/index.tsx
