+++ site/src/components/Report/StatsDisplay/index.tsx
