+++ site/src/components/Report/index.tsx
@@ -1,17 +1,21 @@
 import React from "react";
-import { decreesStats, totalDeptStats } from "../../engine/engine";
+import {
+  DecreesStats,
+  TotalDepartments,
+  TotalMinistrySeries,
+} from "../../engine/engine";
 import SeriesChart, { ministrySeriesDataElement } from "./SeriesChart";
 import s from "./Report.module.scss";
 import Summary from "./Summary";
 import { DepartmentsChart, departmentsDataElement } from "./DepartmentsChart";
 
 interface ReportProps {
-  decreesStats: decreesStats;
+  decreesStats: DecreesStats;
 }
 
 const Report: React.FC<ReportProps> = ({ decreesStats }) => {
   const seriesData = getMinistrySeriesData(
-    decreesStats.totalMinistryNumberStats
+    decreesStats.totalStats.totalMinistrySeriesStats
   );
   return (
     <div className={s.root}>
@@ -21,45 +25,65 @@ const Report: React.FC<ReportProps> = ({ decreesStats }) => {
         data={seriesData}
       />
       <DepartmentsChart
-        data={getDepartmentsData(decreesStats.totalDepartmentsStats)}
+        data={getDepartmentsData(decreesStats.totalStats.totalDepartmentsStats)}
       />
     </div>
   );
 };
 
-function getDecreesNumbers(stats: decreesStats): string[] {
+function getDecreesNumbers(stats: DecreesStats): string[] {
   return stats.decrees.map((d) => d.number);
 }
 
 function getMinistrySeriesData(
-  stats: [string, number][]
+  stats: TotalMinistrySeries[]
 ): ministrySeriesDataElement[] {
   let statsCopy = stats.slice();
   const seriesStatsDataLen = 11;
+  let sortBySeriesFn = (
+    a: TotalMinistrySeries,
+    b: TotalMinistrySeries,
+    asc: boolean = true
+  ) => {
+    let aYear = parseInt(a.ministrySeries.year, 10),
+      bYear = parseInt(b.ministrySeries.year),
+      aSeries = parseInt(a.ministrySeries.series, 10),
+      bSeries = parseInt(b.ministrySeries.series);
+    let result =
+      aYear > bYear ? 1 : aYear < bYear ? -1 : aSeries > bSeries ? 1 : -1;
+    if (!asc) {
+      result *= -1;
+    }
+    return result;
+  };
   if (stats.length > seriesStatsDataLen) {
     statsCopy.sort((a, b) =>
-      a[1] > b[1] ? -1 : a[1] < b[1] ? 1 : a[0] > b[0] ? -1 : 1
+      a.count > b.count
+        ? -1
+        : a.count < b.count
+        ? 1
+        : sortBySeriesFn(a, b, false)
     );
   }
   return ministryNumberStatsToMinistrySeriesDataElements(
-    statsCopy
-      .slice(0, seriesStatsDataLen)
-      .sort((a, b) => (a[0] > b[0] ? 1 : -1))
+    statsCopy.slice(0, seriesStatsDataLen).sort((a, b) => sortBySeriesFn(a, b))
   );
 }
 
 function ministryNumberStatsToMinistrySeriesDataElements(
-  mn: [string, number][]
+  mn: TotalMinistrySeries[]
 ): ministrySeriesDataElement[] {
   return mn.map((m) => {
     return {
-      series: m[0],
-      count: m[1],
+      series: `${m.ministrySeries.year}X ${m.ministrySeries.series}`,
+      count: m.count,
     };
   });
 }
 
-function getDepartmentsData(stats: totalDeptStats[]): departmentsDataElement[] {
+function getDepartmentsData(
+  stats: TotalDepartments[]
+): departmentsDataElement[] {
   let statsCopy = stats.slice();
   const departmentsDataLength = 10;
   statsCopy.sort((a, b) =>
@@ -67,7 +91,7 @@ function getDepartmentsData(stats: totalDeptStats[]): departmentsDataElement[] {
       ? -1
       : a.count < b.count
       ? 1
-      : a.department.idx > b.department.idx
+      : a.department.number > b.department.number
       ? 1
       : -1
   );
@@ -77,10 +101,10 @@ function getDepartmentsData(stats: totalDeptStats[]): departmentsDataElement[] {
 }
 
 function departmentsStatsToDepartmentsDataElements(
-  stats: totalDeptStats[]
+  stats: TotalDepartments[]
 ): departmentsDataElement[] {
   return stats.map((s) => ({
-    idx: s.department.idx,
+    idx: s.department.number,
     name: s.department.name,
     count: s.count,
   }));
